Change: Separate components and containers
